chore: add meta tags to pages
